refactor: move troublesome helper to its single place of use
Having it live with the other helpers causes a recursive loop in
a future commit, because the helpers file will soon need to import
the config file to pull in a variable. Moving this one helper out to its
one place of use fixes the problem.
