+++ server/config.js
@@ -1,5 +1,25 @@
 // @ts-nocheck
-import { assertNoUndefinedValues } from './helpers.js';
+
+/**
+ * @param {object} object
+ */
+export const assertNoUndefinedValues = (object) => {
+	const keysWithUndefinedValues = [];
+
+	for (const [key, value] of Object.entries(object)) {
+		if (typeof value === 'undefined') {
+			keysWithUndefinedValues.push(key);
+		}
+	}
+
+	if (keysWithUndefinedValues.length) {
+		throw new Error(
+			`Expected all config variables to be defined but found the following undefined variables: ${keysWithUndefinedValues.join(
+				', '
+			)}.`
+		);
+	}
+};
 
 const config = {
 	APP_FRIENDLY_NAME: 'fill-out--your-user-facing-app-name',
